// app/types/form.ts
export type FormData = {
  email: string;
  salutation: string;
  firstname: string;
  lastname: string;
  address: string;
  postalCode: string;
  city: string;
  country: string;
  phone: string;
  mobile: string;
  certEmail: string;
  certSalutation: string;
  certFirstname: string;
  newsletter: boolean;
};